task 0

+++ 0-bubble_sort.c
@@ -8,7 +8,7 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	int i, j, temp, k;  /* Declare k here with i, j, temp */
+	size_t i, j, temp, k;  /* Declare k here with i, j, temp */
 
 	for (i = 0; i < size - 1; i++)
 	{
